+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/pushers/PushersService.kt
@@ -17,6 +17,7 @@ package im.vector.matrix.android.api.session.pushers
 
 import androidx.lifecycle.LiveData
 import im.vector.matrix.android.api.MatrixCallback
+import im.vector.matrix.android.api.util.Cancelable
 import java.util.UUID
 
 interface PushersService {
@@ -28,19 +29,32 @@ interface PushersService {
 
     /**
      * Add a new HTTP pusher.
+     * Note that only `http` kind is supported by the SDK for now.
+     * Ref: https://matrix.org/docs/spec/client_server/latest#post-matrix-client-r0-pushers-set
      *
-     * @param pushkey           the pushkey
+     * @param pushkey           This is a unique identifier for this pusher. The value you should use for
+     *                          this is the routing or destination address information for the notification,
+     *                          for example, the APNS token for APNS or the Registration ID for GCM. If your
+     *                          notification client has no such concept, use any unique identifier. Max length, 512 chars.
+     *                          If the kind is "email", this is the email address to send notifications to.
      * @param appId             the application id
-     * @param profileTag        the profile tag
-     * @param lang              the language
-     * @param appDisplayName    a human-readable application name
-     * @param deviceDisplayName a human-readable device name
-     * @param url               the URL that should be used to send notifications
-     * @param append            append the pusher
-     * @param withEventIdOnly   true to limit the push content
+     *                          This is a reverse-DNS style identifier for the application. It is recommended
+     *                          that this end with the platform, such that different platform versions get
+     *                          different app identifiers. Max length, 64 chars.
+     * @param profileTag        This string determines which set of device specific rules this pusher executes.
+     * @param lang              The preferred language for receiving notifications (e.g. "en" or "en-US").
+     * @param appDisplayName    A human readable string that will allow the user to identify what application owns this pusher.
+     * @param deviceDisplayName A human readable string that will allow the user to identify what device owns this pusher.
+     * @param url               The URL to use to send notifications to. MUST be an HTTPS URL with a path of /_matrix/push/v1/notify.
+     * @param append            If true, the homeserver should add another pusher with the given pushkey and App ID in addition
+     *                          to any others with different user IDs. Otherwise, the homeserver must remove any other pushers
+     *                          with the same App ID and pushkey for different users.
+     * @param withEventIdOnly   true to limit the push content to only id and not message content
+     *                          Ref: https://matrix.org/docs/spec/push_gateway/r0.1.1#homeserver-behaviour
      *
-     * @return A work request uuid. Can be used to listen to the status
-     * (LiveData<WorkInfo> status = workManager.getWorkInfoByIdLiveData(<UUID>))
+     * @return                  A work request uuid. Can be used to listen to the status
+     *                          (LiveData<WorkInfo> status = workManager.getWorkInfoByIdLiveData(<UUID>))
+     * @throws [InvalidParameterException] if a parameter is not correct
      */
     fun addHttpPusher(pushkey: String,
                       appId: String,
@@ -52,13 +66,18 @@ interface PushersService {
                       append: Boolean,
                       withEventIdOnly: Boolean): UUID
 
-    fun removeHttpPusher(pushkey: String, appId: String, callback: MatrixCallback<Unit>)
+    /**
+     * Remove the http pusher
+     */
+    fun removeHttpPusher(pushkey: String, appId: String, callback: MatrixCallback<Unit>): Cancelable
 
-    companion object {
-        const val EVENT_ID_ONLY = "event_id_only"
-    }
-
+    /**
+     * Get the current pushers, as a LiveData
+     */
     fun getPushersLive(): LiveData<List<Pusher>>
 
-    fun pushers() : List<Pusher>
+    /**
+     * Get the current pushers
+     */
+    fun getPushers(): List<Pusher>
 }