+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/api/session/room/model/relation/RelationDefaultContent.kt
@@ -22,5 +22,6 @@ import com.squareup.moshi.JsonClass
 data class RelationDefaultContent(
         @Json(name = "rel_type") override val type: String?,
         @Json(name = "event_id") override val eventId: String?,
-        @Json(name = "m.in_reply_to") override val inReplyTo: ReplyToContent? = null
+        @Json(name = "m.in_reply_to") override val inReplyTo: ReplyToContent? = null,
+        @Json(name = "option") override val option: Int? = null
 ) : RelationContent